+++ package.json
@@ -7,6 +7,7 @@
   "types": "index.d.ts",
   "scripts": {
     "dev": "vite",
+    "pack": "node scripts/publish.js",
     "build": "node scripts/build.js",
     "preview": "vite preview",
     "lint": "eslint \"**/*.{js,ts}\"",
@@ -23,10 +24,7 @@
   "dependencies": {
     "@shoelace-style/shoelace": "^2.11.2",
     "chromatic": "^10.1.0",
-    "cpy": "^11.0.0",
-    "cpy-cli": "^5.0.0",
-    "lit": "^3.0.2",
-    "style-dictionary-utils": "^2.0.4"
+    "lit": "^3.0.2"
   },
   "devDependencies": {
     "@mdx-js/loader": "^3.0.0",
@@ -38,6 +36,8 @@
     "@storybook/web-components-vite": "^7.6.4",
     "@typescript-eslint/eslint-plugin": "^6.12.0",
     "@typescript-eslint/parser": "^6.12.0",
+    "cpy": "^11.0.0",
+    "cpy-cli": "^5.0.0",
     "eslint": "^8.54.0",
     "eslint-config-prettier": "^9.0.0",
     "eslint-plugin-lit": "^1.10.1",
@@ -46,8 +46,10 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "storybook": "^7.6.4",
+    "style-dictionary-utils": "^2.0.4",
     "typescript": "^5.2.2",
     "vite": "^5.0.0",
+    "vite-plugin-css-injected-by-js": "^3.3.1",
     "vite-plugin-dts": "^3.7.0",
     "web-component-analyzer": "^2.0.0"
   },
