+++ scripts/nextTask.js
@@ -1,0 +1,20 @@
+// Helper function to run tasks in async
+import ora from 'ora';
+import chalk from 'chalk';
+const spinner = ora({ hideCursor: false }).start();
+export default async function nextTask(label, action) {
+  spinner.text = label;
+  spinner.start();
+
+  try {
+    await action();
+    spinner.stop();
+    console.log(`${chalk.green('✔')} ${label}`);
+  } catch (err) {
+    spinner.stop();
+    console.error(`${chalk.red('✘')} ${err}`);
+    if (err.stdout) console.error(chalk.red(err.stdout));
+    if (err.stderr) console.error(chalk.red(err.stderr));
+    process.exit(1);
+  }
+}
