+++ vite.config.ts
@@ -2,9 +2,11 @@
 import { defineConfig } from 'vite';
 import { resolve } from 'path';
 import dts from 'vite-plugin-dts';
+import cssInjectedByJsPlugin from 'vite-plugin-css-injected-by-js';
 
 export default defineConfig({
   plugins: [
+    cssInjectedByJsPlugin(),
     dts({
       include: ['src/index.ts'],
     }),
